Refactor config service modules
