+++ src/zivo/services/config/render.py
@@ -1,0 +1,150 @@
+"""Render the normalized application config back to TOML."""
+
+from __future__ import annotations
+
+from zivo.models import AppConfig
+
+from .shared import HELP_BAR_FIELDS
+
+
+def render_default_config() -> str:
+    """Render the default config file contents."""
+
+    return render_app_config(AppConfig())
+
+
+def render_app_config(config: AppConfig) -> str:
+    """Render the normalized application config as TOML."""
+
+    sections = [
+        render_terminal_section(config),
+        render_editor_section(config),
+        render_display_section(config),
+        render_behavior_section(config),
+        render_logging_section(config),
+        render_bookmarks_section(config),
+        render_help_bar_section(config),
+    ]
+    return "\n\n".join(sections) + "\n"
+
+
+def render_terminal_section(config: AppConfig) -> str:
+    linux = render_command_array(config.terminal.linux)
+    macos = render_command_array(config.terminal.macos)
+    windows = render_command_array(config.terminal.windows)
+    return (
+        "[terminal]\n"
+        "# Optional OS-specific terminal launch templates.\n"
+        "# Use {path} for the working directory.\n"
+        "# Examples:\n"
+        '# linux = [\n'
+        '#   "konsole --working-directory {path}",\n'
+        '#   "gnome-terminal --working-directory={path}",\n'
+        '# ]\n'
+        '# macos = ["open -a Terminal {path}"]\n'
+        '# windows = ["wt -d {path}"]\n'
+        f"linux = [{linux}]\n"
+        f"macos = [{macos}]\n"
+        f"windows = [{windows}]"
+    )
+
+
+def render_editor_section(config: AppConfig) -> str:
+    command = render_optional_toml_string(config.editor.command)
+    return (
+        "[editor]\n"
+        "# Optional terminal editor command for `e`.\n"
+        "# Use a shell-style command without the file path; zivo appends it automatically.\n"
+        "# Examples:\n"
+        '# command = "nvim -u NONE"\n'
+        '# command = "emacs -nw"\n'
+        f"command = {command}"
+    )
+
+
+def render_display_section(config: AppConfig) -> str:
+    return (
+        "[display]\n"
+        f"show_hidden_files = {render_bool(config.display.show_hidden_files)}\n"
+        f"show_directory_sizes = {render_bool(config.display.show_directory_sizes)}\n"
+        f"show_preview = {render_bool(config.display.show_preview)}\n"
+        f'theme = "{config.display.theme}"\n'
+        f'preview_syntax_theme = "{config.display.preview_syntax_theme}"\n'
+        f"preview_max_kib = {config.display.preview_max_kib}\n"
+        f'default_sort_field = "{config.display.default_sort_field}"\n'
+        f"default_sort_descending = {render_bool(config.display.default_sort_descending)}\n"
+        f"directories_first = {render_bool(config.display.directories_first)}\n"
+        f"grep_preview_context_lines = {config.display.grep_preview_context_lines}\n"
+        f'split_terminal_position = "{config.display.split_terminal_position}"'
+    )
+
+
+def render_behavior_section(config: AppConfig) -> str:
+    return (
+        "[behavior]\n"
+        f"confirm_delete = {render_bool(config.behavior.confirm_delete)}\n"
+        f'paste_conflict_action = "{config.behavior.paste_conflict_action}"'
+    )
+
+
+def render_logging_section(config: AppConfig) -> str:
+    path = render_optional_toml_string(config.logging.path)
+    return (
+        "[logging]\n"
+        "# Optional file output for startup and unhandled exceptions.\n"
+        "# Leave empty to write zivo.log next to config.toml.\n"
+        f"enabled = {render_bool(config.logging.enabled)}\n"
+        f'level = "{config.logging.level}"\n'
+        f"path = {path}"
+    )
+
+
+def render_bookmarks_section(config: AppConfig) -> str:
+    paths = render_command_array(config.bookmarks.paths)
+    return (
+        "[bookmarks]\n"
+        "# Optional bookmarked directories shown in the command palette.\n"
+        "# Use absolute paths.\n"
+        "# Example:\n"
+        '# paths = ["/home/user/src", "/home/user/docs"]\n'
+        f"paths = [{paths}]"
+    )
+
+
+def render_help_bar_section(config: AppConfig) -> str:
+    lines = [
+        "[help_bar]",
+        "# Optional custom help bar text for each UI mode.",
+        "# Leave empty to use built-in defaults.",
+        "# Example:",
+        '# browsing = ["Custom help line 1", "Custom help line 2"]',
+    ]
+    for field in HELP_BAR_FIELDS:
+        lines.append(f"{field} = {render_help_lines(getattr(config.help_bar, field))}")
+    return "\n".join(lines)
+
+
+def render_command_array(commands: tuple[str, ...]) -> str:
+    return ", ".join(render_toml_string(command) for command in commands)
+
+
+def render_bool(value: bool) -> str:
+    return "true" if value else "false"
+
+
+def render_toml_string(value: str) -> str:
+    escaped = value.replace("\\", "\\\\").replace('"', '\\"')
+    return f'"{escaped}"'
+
+
+def render_optional_toml_string(value: str | None) -> str:
+    if value is None:
+        return '""'
+    return render_toml_string(value)
+
+
+def render_help_lines(lines: tuple[str, ...]) -> str:
+    if not lines:
+        return "[]"
+    rendered = ", ".join(render_toml_string(line) for line in lines)
+    return f"[{rendered}]"
